api/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==22.0.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.13.2

# AWS
boto3==1.35.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# AI/LLM
anthropic==0.34.0
google-genai==1.0.0

# PDF Processing
pymupdf==1.24.10
pymupdf4llm==0.0.17

# Data
pandas==2.2.2
openpyxl==3.1.5
xlsxwriter==3.2.0
pydantic==2.9.1

# Utilities
python-dotenv==1.0.1
aiohttp==3.10.5
aiofiles==24.1.0
httpx==0.27.0
